+++ docs/Changes.md
@@ -16,6 +16,7 @@ This document presents the important changes made to each KioskForge release fro
 4. The Wi-Fi password is now encoded on the kiosk so that hackers cannot read it in plain text anymore (a security precaution).  This made the crude wipe of the network-data file on the installation medium redundant so that it has been removed.
 5. Yet another attempt was made to make Ubuntu Server pick up the "regulatory domain" (Wi-Fi country code) at every boot.  This is not easy, though, because Ubuntu Server for Pi does not provide any working way to set this value.  As far as I can see from posts on the internet, the Linux kernel asks the Access Point (AP) for this information.  Most of the time, this works, but sometimes it doesn't.  There is not anything I can do about it to my knowledge.
 6. The kiosk hostname generated if the `hostname` option is empty now includes a separating dash between `kiosk` and the number: `kiosk-34271`.  It is strongly recommended that you specify a hostname, though, as this makes it easy to see the purpose of the kiosk in smarter routers.
+7. The kiosk forge process now waits indefinitely until an online network connection is detected so that the end-user has plenty of time to identify and fix the issue.
 
 ## Version 1.03 (2026.05.27)
 1. Ubuntu Server 26.04.x is now supported by KioskForge.  This is beta as the tests have been limited (everything worked right away in these tests, though).